+++ libraries/scrcpy/src/1_15/impl/index.ts
@@ -1,5 +1,6 @@
 export * from "./back-or-screen-on.js";
 export * from "./clipboard-stream.js";
+export * from "./codec-options.js";
 export * from "./control-message-types.js";
 export * from "./defaults.js";
 export * from "./init.js";
